+++ lib/action_operation.rb
@@ -11,8 +11,8 @@ module ActionOperation
 
   State = Struct.new(:raw)
   Drift = Struct.new(:to)
-  Task = Struct.new(:name, :required)
-  Catch = Struct.new(:name, :exception)
+  Task = Struct.new(:name, :receiver, :required)
+  Catch = Struct.new(:name, :receiver, :exception)
 
   def initialize(raw:)
     raise ArgumentError, "needs to be a Hash" unless raw.kind_of?(Hash)
@@ -20,31 +20,72 @@ module ActionOperation
     @raw = raw
   end
 
+  private def callbackings(arounds, &process)
+    arounds.reverse.reduce(process) do |compound, callback|
+      callback.call(&compound).call
+    end.call
+  end
+
   def call(start: nil, raw: @raw)
+    around_steps do
+      begin
+        around_tasks do
+          tasks(start, raw)
+        end
+      rescue *left.select(&:exception).map(&:exception).uniq => handled_exception
+        around_catches do
+          catches(handled_exception, raw)
+        end
+      end
+    end
+  end
+
+  private def tasks(start, raw)
     right.from(start || 0).reduce(raw) do |state, step|
       next state unless step.required || (start && right.at(start) == step)
 
-      raise Error::MissingTask, step unless respond_to?(step.name)
-      raise Error::MissingSchema, step unless self.class.schemas.key?(step.name)
+      raise Error::MissingTask, step: step unless respond_to?(step.name)
+      raise Error::MissingSchema, step: step unless self.class.schemas.key?(step.name)
 
-      # NOTE: We only care about this so we can refernece it in the rescue
+      # NOTE: We only care about this so we can reference it in the rescue
       @latest_step = step
 
-      value = public_send(step.name, state: self.class.schemas.fetch(step.name).new(state))
+      # puts "#{step.class}::#{step.receiver}##{step.name}"
+
+      begin
+        value = around_task do
+          public_send(step.name, state: self.class.schemas.fetch(step.name).new(state))
+        end
+        # puts "#{step.class}::#{step.receiver}##{step.name} #{value}"
+      rescue SmartParams::Error::InvalidPropertyType => invalid_property_type_exception
+        raise Error::StepSchemaMismatch, step: step, schema: self.class.schemas.fetch(step.name), raw: raw, cause: invalid_property_type_exception
+      end
+
 
       case value
-      when State then value.raw
-      when Drift then break call(start: right.find_index { |step| step.name == value.to }, raw: raw)
-      else state
+        when State then value.raw
+        when Drift then break call(start: right.find_index { |step| step.name == value.to }, raw: state)
+        else state
       end
     end
-  rescue *left.select(&:exception).map(&:exception).uniq => handled_exception
-    left.select do |failure|
-      failure.exception === handled_exception
-    end.reduce(handled_exception) do |exception, step|
-      raise Error::MissingError, step unless respond_to?(step.name)
+  end
 
-      value = public_send(step.name, exception: exception, state: self.class.schemas.fetch(@latest_step.name).new(raw), step: @latest_step)
+  private def catches(exception, raw)
+    left.select do |failure|
+      failure.exception === exception
+    end.reduce(exception) do |exception, step|
+      raise Error::MissingError, step: step unless respond_to?(step.name)
+
+      # puts "#{step.class}::#{step.receiver}##{step.name}"
+
+      begin
+        value = around_catch do
+          public_send(step.name, exception: exception, state: raw, step: @latest_step)
+        end
+        # puts "#{step.class}::#{step.receiver}##{step.name} #{value}"
+      rescue SmartParams::Error::InvalidPropertyType => invalid_property_type_exception
+        raise Error::StepSchemaMismatch, step: @latest_step, schema: self.class.schemas.fetch(@latest_step.name), raw: raw, cause: invalid_property_type_exception
+      end
 
       if value.kind_of?(Drift)
         break call(start: right.find_index { |step| step.name == value.to }, raw: raw)
@@ -64,6 +105,30 @@ module ActionOperation
     raise ArgumentError, "needs to be a Symbol or String" unless to.kind_of?(Symbol) || to.kind_of?(String)
 
     Drift.new(to)
+  end
+
+  def around_steps(&callback)
+    callback.call
+  end
+
+  def around_step(&callback)
+    callback.call
+  end
+
+  def around_tasks(&callback)
+    callback.call
+  end
+
+  def around_task(&callback)
+    callback.call
+  end
+
+  def around_catches(&callback)
+    callback.call
+  end
+
+  def around_catch(&callback)
+    callback.call
   end
 
   private def left
@@ -92,11 +157,11 @@ module ActionOperation
     end
 
     def task(name, required: true)
-      right << Task.new(name, required)
+      right << Task.new(name, self, required)
     end
 
     def catch(name, exception: StandardError)
-      left << Catch.new(name, exception)
+      left << Catch.new(name, self, exception)
     end
 
     def right