+++ lib/action_operation/error.rb
@@ -3,5 +3,6 @@ module ActionOperation
     require_relative "error/missing_error"
     require_relative "error/missing_schema"
     require_relative "error/missing_task"
+    require_relative "error/step_schema_mismatch"
   end
 end